Reorganize/consolidate test modules
