+++ tests/test_traversal.py
@@ -4,18 +4,15 @@
 
 """Tests for the traversal module."""
 
-import platform
 from argparse import Namespace
 from pathlib import Path
 from tempfile import TemporaryDirectory
 from unittest.mock import call, patch
 
-import pytest
 from conftest import SymlinkMock
 
 import pyclean.main
 from pyclean.bytecode import BYTECODE_DIRS, BYTECODE_FILES
-from pyclean.ignore import normalize, should_ignore
 from pyclean.traversal import descend_and_clean
 
 
@@ -53,91 +50,6 @@ def test_skip_ignored_directories(mock_log):
     descend_and_clean(directory, BYTECODE_FILES, BYTECODE_DIRS)
 
     mock_log.debug.assert_called_with('Skipping %s', '.git')
-
-
-@pytest.mark.parametrize(
-    ('path_str', 'patterns', 'expected'),
-    [
-        # Simple name matches
-        ('foo/bar', ['bar'], True),
-        ('foo/baz/bar', ['bar'], True),
-        ('bar', ['bar'], True),
-        ('foo/bar', ['baz'], False),
-        # Path matches
-        ('foo/bar', ['foo/bar'], True),
-        ('baz/foo/bar', ['foo/bar'], True),
-        ('test/foo/bar', ['foo/bar'], True),
-        ('foo/bar/baz', ['foo/bar'], True),  # Subdirectories are also ignored
-        ('bar/foo', ['foo/bar'], False),
-        ('foo/baz', ['foo/bar'], False),
-        # Multiple patterns
-        ('foo/bar', ['baz', 'bar'], True),
-        ('foo/bar', ['baz', 'foo/bar'], True),
-        ('foo/bar', ['test', 'data'], False),
-        # Edge cases
-        ('bar', ['foo/bar'], False),
-        ('foo', ['foo/bar'], False),
-        # Pattern longer than path
-        ('baz', ['foo/bar/baz'], False),
-        ('bar/baz', ['foo/bar/baz'], False),
-        # None/empty patterns
-        ('foo/bar', None, False),
-        ('foo/bar', [], False),
-        # Subdirectories of ignored paths
-        ('foo/bar/baz/deep', ['foo/bar'], True),
-        ('src/foo/bar/models', ['foo/bar'], True),
-        ('foo/bar/baz', ['foo/bar/baz'], True),
-    ],
-)
-def test_should_ignore(path_str, patterns, expected):
-    """
-    Does should_ignore correctly match path patterns?
-    """
-    result = should_ignore(path_str, patterns)
-    assert result == expected
-
-
-@pytest.mark.skipif(platform.system() != 'Windows', reason='Windows-specific test')
-@pytest.mark.parametrize(
-    ('path_str', 'patterns', 'expected'),
-    [
-        # Windows-style pattern (backslash) matching filesystem paths
-        ('foo/bar', [r'foo\bar'], True),
-        ('foo/bar/baz', [r'foo\bar'], True),
-        ('src/foo/bar', [r'foo\bar'], True),
-    ],
-)
-def test_should_ignore_windows_paths(path_str, patterns, expected):
-    """
-    Does should_ignore correctly handle Windows-style backslash patterns?
-    This test only runs on Windows where backslash is a path separator.
-    """
-    result = should_ignore(path_str, patterns)
-    assert result == expected
-
-
-@pytest.mark.skipif(
-    platform.system() not in ['Linux', 'Darwin'],
-    reason='Unix-specific test',
-)
-def test_normalize_pattern_posix():
-    """
-    Does normalize preserve patterns on Unix?
-    On Unix, backslash can be part of a filename.
-    """
-    assert normalize('foo/bar') == 'foo/bar'
-    assert normalize(r'foo\bar') == r'foo\bar'  # Preserved on Linux and macOS
-    assert normalize('bar') == 'bar'
-
-
-@pytest.mark.skipif(platform.system() != 'Windows', reason='Windows-specific test')
-def test_normalize_pattern_windows():
-    """
-    Does normalize convert backslashes to forward slashes on Windows?
-    """
-    assert normalize('foo/bar') == 'foo/bar'
-    assert normalize(r'foo\bar') == 'foo/bar'  # Normalized
-    assert normalize('bar') == 'bar'
 
 
 def test_ignore_with_simple_name():
